+++ core-business/core-user/user-web/src/main/java/com/xzn/user/service/Impl/BaseServiceImpl.java
@@ -28,7 +28,7 @@ public class BaseServiceImpl<M extends BaseMapper<T>, T> implements BaseService<
     }
 
     @Override
-    public T selectById(Integer id) {
+    public T selectById(long id) {
         return mapper.selectById(id);
     }
